% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeCSSFile.R
\name{includeCSSFile}
\alias{includeCSSFile}
\title{Include CSS as a file}
\usage{
includeCSSFile(path)
}
\arguments{
\item{path}{Local path or URL to the css file to link}
}
\description{
Similar to `shiny::includeCSS` but inserts the css as a linked file instead
of inserting the content directly. This is more convenient to track css
properties using a browser's developer tools because the CSS source is more
obvious.
}
