% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhelpr-package.R
\docType{package}
\name{shinyhelpr-package}
\alias{shinyhelpr}
\alias{shinyhelpr-package}
\title{shinyhelpr: Helper functions for shiny}
\description{
Misc. helper functions for shiny.
}
\author{
\strong{Maintainer}: Antoine Sachet \email{antoine@soundout.com}

Other contributors:
\itemize{
  \item SoundOut Ltd [copyright holder]
}

}
\keyword{internal}
