% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_favicon.R
\name{use_favicon}
\alias{use_favicon}
\title{Add a favicon to a shiny app}
\usage{
use_favicon(favicon)
}
\arguments{
\item{favicon}{Path (local or URL) to the favicon file.}
}
\description{
Call `use_favicon` anywhere in your shiny app to add the specified favicon.
}
\examples{

\dontrun{
eye_logo <- paste0("https://thegraphicsfairy.com/wp-content/",
  "uploads/2013/10/Free-Public-Domain-Watching-Eye-Image-GraphicsFairy.jpg")
# Blank page with a favicon
shinyApp(ui = fluidPage(use_favicon(eye_logo)),
         server = function(input, output) {})
}
}
